-*- Fixed map2cs so that it uses the portable strcasecmp() rather than the     unportable stricmp().
-*- Eliminated a compilation warning in demoop.cpp caused by mismatch of
    printf() format directive and incoming argument.


git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@6744 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ apps/demo/demoop.cpp
@@ -195,7 +195,7 @@ void HideMeshOp::Do (cs_time /*dt*/)
 void TestOp::Do (cs_time dt)
 {
   printf ("dt=%ld fps=%g\n", dt,
-  	DemoSequenceManager::demoseq->GetFPS ()); fflush (stdout);
+  	(long)DemoSequenceManager::demoseq->GetFPS ()); fflush (stdout);
 }
 
 //-----------------------------------------------------------------------------
